Two-letter abbreviations for octant circumplex scales

Return a vector of abbreviations for octant circumplex scales, from PA to NO.

Usage

PANO()

Value

A character vector with eight elements, each corresponding to the abbreviation of an octant subscale: PA, BC, DE, FG, HI, JK, LM, NO.

Examples

PANO()
#> [1] "PA" "BC" "DE" "FG" "HI" "JK" "LM" "NO"