Test circumplex fit using the RANDALL Concordance Index

Conduct the Randomization Test of Hypothesized Order Relations (RANDALL). Estimate the RANDALL Concordance Index using a Monte-Carlo-based bootstrap procedure and estimate its percentile confidence interval.

fit_randall(.data, scales, boots = 2000, interval = 0.95,
  listwise = TRUE, tolerance = 0.1)

Arguments

.data
Required. A data frame containing at least circumplex scales.

scales
Required. The variable names or column numbers for the variables in .data that contain circumplex scales to be analyzed.

boots
Optional. A single positive integer indicating how many bootstrap resamples to use when estimating the confidence intervals (default = 2000).

interval
Optional. A single positive number between 0 and 1 (exclusive) that indicates what confidence level to use when estimating the confidence intervals (default = 0.95).

listwise
Optional. A logical indicating whether missing values should be handled by listwise deletion (TRUE) or pairwise deletion (FALSE). Note that pairwise deletion may result in different missing data patterns in each bootstrap resample and is slower to compute (default = TRUE).

tolerance
Optional. Tolerance (relative to largest variance) for numerical lack of positive-definiteness in the correlation matrix computed among scales (default = 0.1). See MASS::mvrnorm.

Value

A three-element vector containing the mean concordance index, the lower bound of the confidence interval, and the upper bound of the confidence interval. (Will soon be replaced by a list object.)

See also

Examples

data("jz2017")
fit_randall(.data = jz2017, scales = PA:NO)

#> [1] 0.8055556 0.7569444 0.8402778