Inventory of Interpersonal Problems, Brief Version

Information about the Inventory of Interpersonal Problems, Brief Version (IIP-32). Note that, although we have permission to provide some information about the IIP-32, Mind Garden Inc. has exclusive rights to distribute it in full.

iip32

Format

An object of class circumplex_instrument of length 5.

Source

Horowitz, L. M., Alden, L. E., Wiggins, J. S., & Pincus, A. L. (2000). IIP-64/IIP-32 professional manual. San Antonio, TX: The Psychological Corporation.

https://www.mindgarden.com/113-inventory-of-interpersonal-problems

Examples

instrument("iip32")
summary(iip32)

#> IIP-32: Inventory of Interpersonal Problems, Brief Version
#> 32 items, 8 scales, 3 normative data sets
#> Horowitz, Alden, Wiggins, & Pincus (2003)
#> <https://www.mindgarden.com/113-inventory-of-interpersonal-problems>
#> 
#> The IIP-32 contains 8 circumplex scales.
#> PA: Domineering/Controlling (90 degrees)
#> BC: Vindictive/Self-Centered (135 degrees)
#> DE: Cold/Distant (180 degrees)
#> FG: Socially Inhibited (225 degrees)
#> HI: Nonassertive (270 degrees)
#> JK: Overly Accommodating (315 degrees)
#> LM: Self-Sacrificing (360 degrees)
#> NO: Intrusive/Needy (45 degrees)
#> 
#> The IIP-32 is rated using the following 5-point scale.
#> 0. Not at all
#> 1. A little bit
#> 2. Moderately
#> 3. Quite a bit
#> 4. Extremely
#> 
#> The IIP-32 contains 32 items (copyrighted):
#> Visit the Mind Garden Inc. website for item text and numbering.
#> 
#> The IIP-32 currently has 3 normative data set(s):
#> 1. 800 American community adults, overall
#> Horowitz, Alden, Wiggins, & Pincus (2003)
#> <https://www.mindgarden.com/113-inventory-of-interpersonal-problems>
#> 2. 400 American community adults, females
#> Horowitz, Alden, Wiggins, & Pincus (2003)
#> <https://www.mindgarden.com/113-inventory-of-interpersonal-problems>
#> 3. 400 American community adults, males
#> Horowitz, Alden, Wiggins, & Pincus (2003)
#> <https://www.mindgarden.com/113-inventory-of-interpersonal-problems>