Inventory of Interpersonal Problems, Short Circumplex

Information about the Inventory of Interpersonal Problems Short Circumplex (IIP-SC).

Usage

iipsc

Format

An object of class circumplex_instrument of length 5.

Source

Soldz, S., Budman, S., Demby, A., & Merry, J. (1995). A short form of the Inventory of Interpersonal Problems Circumplex Scales. Assessment, 2(1), 53-63.

doi:10.1080/00223890802388665

Examples

instrument("iipsc")
summary(iipsc)
#> IIP-SC: Inventory of Interpersonal Problems Short Circumplex
#> 32 items, 8 scales, 2 normative data sets
#> Soldz, Budman, Demby, & Merry (1995)
#> <https://doi.org/10.1177/1073191195002001006>
#> 
#> The IIP-SC contains 8 circumplex scales.
#> PA: Domineering (90 degrees)
#> BC: Vindictive (135 degrees)
#> DE: Cold (180 degrees)
#> FG: Socially Avoidant (225 degrees)
#> HI: Nonassertive (270 degrees)
#> JK: Exploitable (315 degrees)
#> LM: Overly Nurturant (360 degrees)
#> NO: Intrusive (45 degrees)
#> 
#> The IIP-SC is rated using the following 5-point scale.
#> 0. Not at all
#> 1. Somewhat
#> 2. Moderately
#> 3. Very
#> 4. Extremely
#> 
#> The IIP-SC contains 32 items (partial text):
#> 1. ...point of view...
#> 2. ...supportive of another...
#> 3. ...show affection to...
#> 4. ...join in on...
#> 5. ...stop bothering me...
#> 6. ...I am angry...
#> 7. ...my own welfare...
#> 8. ...keep things private...
#> 9. ...too aggressive toward...
#> 10. ...another person's happiness...
#> 11. ...feeling of love...
#> 12. ...introduce myself to...
#> 13. ...confront people with...
#> 14. ...assertive without worrying...
#> 15. ...please other people...
#> 16. ...open up to...
#> 17. ...control other people...
#> 18. ...too suspicious of...
#> 19. ...feel close to...
#> 20. ...socialize with other...
#> 21. ...assertive with another...
#> 22. ...too easily persuaded...
#> 23. ...other people's needs...
#> 24. ...noticed too much...
#> 25. ...argue with other...
#> 26. ...revenge against people...
#> 27. ...at a distance...
#> 28. ...get together socially...
#> 29. ...to be firm...
#> 30. ...people take advantage...
#> 31. ...another person's misery...
#> 32. ...tell personal things...
#> 
#> The IIP-SC currently has 2 normative data set(s):
#> 1. 872 American college students
#> Hopwood, Pincus, DeMoor, & Koonce (2011)
#> <https://doi.org/10.1080/00223890802388665>
#> 2. 106 American psychiatric outpatients
#> Soldz, Budman, Demby, & Merry (1995)
#> <https://doi.org/10.1177/1073191195002001006>