Instrument: Interpersonal Sensitivities Circumplex

Information about the Interpersonal Sensitivities Circumplex.

isc

Format

An object of class instrument of length 5.

Source

https://doi.org/10.1111/j.1467-6494.2011.00696.x

Examples

instrument("isc")
summary(isc)

#> ISC: Interpersonal Sensitivities Circumplex
#> 64 items, 8 scales, 1 normative data sets
#> Hopwood et al. (2011)
#> <https://doi.org/10.1111/j.1467-6494.2011.00696.x>
#> 
#> The ISC contains 8 circumplex scales.
#> PA (090 deg): Sensitive to Control 
#> BC (135 deg): Sensitive to Antagonism 
#> DE (180 deg): Sensitive to Remoteness 
#> FG (225 deg): Sensitive to Timidity 
#> HI (270 deg): Sensitive to Passivity 
#> JK (315 deg): Sensitive to Dependence 
#> LM (360 deg): Sensitive to Affection 
#> NO (045 deg): Sensitive to Attention-Seeking 
#> 
#> The ISC is rated using the following 8-point scale. 
#> 1. Not at all, never bothers me 
#> 2. Very little, rarely bothers me 
#> 3. A little, occasionally bothers me 
#> 4. Slightly, bothers me less than half the time 
#> 5. Somewhat, bothers me more than half of the time 
#> 6. Quite, bothers me frequently 
#> 7. Very much, bothers me most of the time 
#> 8. Extremely, always bothers me 
#> 
#> The ISC contains 64 items.
#> 1. Thinks they are my boss 
#> 2. Disregards my feelings 
#> 3. Doesn't show any feelings 
#> 4. Won't engage in conversation 
#> 5. Cannot assert themselves 
#> 6. Avoids conflict at all costs 
#> 7. Wants to spend lots of time with me 
#> 8. Doesn't respect my privacy 
#> 9. Orders me around 
#> 10. Doesn't respond to me 
#> 11. Doesn't want to be friends 
#> 12. Is aloof 
#> 13. Doesn't have a backbone 
#> 14. Always needs support 
#> 15. Tries to get close to me really fast 
#> 16. Talks themselves up 
#> 17. Talks down to me 
#> 18. Is hostile 
#> 19. Walls themselves off from me 
#> 20. Acts like a wallflower 
#> 21. Is weak 
#> 22. Always seems to need my help 
#> 23. Tells me they love me 
#> 24. Shows off 
#> 25. Will do anything to get what they want 
#> 26. Is deceitful 
#> 27. Won't share their feelings with me 
#> 28. Won't step up to the plate 
#> 29. Allows themselves to be dominated by others 
#> 30. Acts helpless 
#> 31. Expresses concern about me 
#> 32. Talks about themselves 
#> 33. Has to call the shots 
#> 34. Mistrusts me 
#> 35. Is unresponsive 
#> 36. Doesn't share their ideas 
#> 37. Is always submissive 
#> 38. Laughs at all my jokes 
#> 39. Acts like we're friends when we don't even know each other 
#> 40. Has to be right 
#> 41. Intimidates me 
#> 42. Doesn't care about my feelings 
#> 43. Avoids me 
#> 44. Won't participate 
#> 45. Cannot make decisions 
#> 46. Is dependent on me 
#> 47. Humors me 
#> 48. Interrupts 
#> 49. Is bossy 
#> 50. Expects the worst out of me 
#> 51. Never gets in touch with me 
#> 52. Is really shy 
#> 53. Is very passive 
#> 54. Believes everything I say 
#> 55. Believes I can do no wrong 
#> 56. Tries to show me how to do things 
#> 57. Always puts themselves first 
#> 58. Is mean-spirited 
#> 59. Pulls away from me 
#> 60. Refuses to lead 
#> 61. Is easily controlled 
#> 62. Is soft 
#> 63. Is clingy 
#> 64. Takes control 
#> 
#> The ISC currently has 1 normative data set(s): 
#> 1. 649 American college students 
#>    Hopwood et al. (2011) 
#>    <https://doi.org/10.1111/j.1467-6494.2011.00696.x>