Angular displacements for quadrant circumplex scales

Return a vector of angular displacements, in degrees, for four equally spaced circumplex scales corresponding to the circumplex quadrants. Can be passed to the angles parameter of other functions in this package.

quadrants()

Value

A numeric vector with eight elements, each corresponding to the angular displacement (in degrees) of a subscale, in the following order: BC, FG, JK, NO.

Examples

quadrants()

#> [1] 135 225 315  45
#> attr(,"class")
#> [1] "circumplex_degree" "numeric"