Standardize circumplex scales using normative data

Take in a data frame containing circumplex scales, angle definitions for each scale, and normative data (from the package or custom) and return that same data frame with each specified circumplex scale transformed into standard scores (i.e., z-scores) based on comparison to the normative data.

standardize(.data, scales, angles, norms)

Arguments

.data
Required. A data frame containing at least circumplex scales.

scales
Required. The variable names or column numbers for the variables in .data that contain circumplex scales to be standardized.

angles
Required. A numeric vector containing the angular displacement of each circumplex scale included in scales (in degrees).

norms
Required. A data frame containing normative data for each circumplex scale included in scales. The following variables are required: Angle, M, and SD. See ?iipsc.

Value

A data frame that matches .data except that the variables included in scales have been transformed into standard scores.

Examples

data("jz2017")
data("iipsc")
standardize(jz2017, scales = PA:NO, angles = octants(), norms = iipsc)

#> # A tibble: 1,166 x 19
#>    Gender      PA     BC     DE      FG      HI     JK     LM      NO PARPD
#>    <fct>    <dbl>  <dbl>  <dbl>   <dbl>   <dbl>  <dbl>  <dbl>   <dbl> <int>
#>  1 Female  1.12    1.03   0.409 -0.0501  0.634   1.31   0.952  1.84       4
#>  2 Female -1.15   -0.786 -1.05  -0.842  -0.186   0.428  0.952  1.53       1
#>  3 Female -1.15   -1.15  -1.05  -1.11   -1.55   -1.62  -1.78  -1.28       0
#>  4 Male    1.88    1.39   0.994  1.53    0.634   0.428  0.648  1.84       1
#>  5 Female -0.773  -0.424 -0.760 -1.11   -1.55   -1.62  -1.78  -1.28       0
#>  6 Male    1.12    1.39   1.58   0.741   0.634  -0.158  0.952  1.84       5
#>  7 Male    1.88    1.39   0.994  1.01    0.0874 -0.158 -0.261  0.906      3
#>  8 Female -1.15   -0.786 -0.760 -0.842  -0.459  -0.158 -0.261 -0.0312     0
#>  9 Female -1.15   -1.15  -1.05  -1.11   -1.28   -1.33  -0.564 -1.28       0
#> 10 Male   -0.0152  1.75   0.994  1.53    0.634   0.135  0.648  1.53       4
#> # ... with 1,156 more rows, and 9 more variables: SCZPD <int>, SZTPD <int>,
#> #   ASPD <int>, BORPD <int>, HISPD <int>, NARPD <int>, AVPD <int>, DPNPD <int>,
#> #   OCPD <int>