Tidy eval helpers

sym() creates a symbol from a string and syms() creates a list of symbols from a character vector.

expr() and quo() quote one expression. quo() wraps the quoted expression in a quosure.

The plural variants exprs() and quos() return a list of quoted expressions or quosures.

enexpr() and enquo() capture the expression supplied as argument by the user of the current function (enquo() wraps this expression in a quosure).

enexprs() and enquos() capture multiple expressions supplied as arguments, including ....

exprs() is not exported to avoid conflicts with Biobase::exprs(), therefore one should always use rlang::exprs().

To learn more about tidy eval and how to use these tools, visit http://rlang.r-lib.org and the Metaprogramming section of Advanced R.