Perform analyses using the Structural Summary Method

Calculate SSM parameters with bootstrapped confidence intervals for a variety of different analysis types. Depending on what arguments are supplied, either mean-based or correlation-based analyses will be performed, one or more groups will be used to stratify the data, and contrasts between groups or measures will be calculated.

ssm_analyze(.data, scales, angles, measures = NULL, grouping = NULL,
  contrast = "none", boots = 2000, interval = 0.95,
  listwise = TRUE)

Arguments

.data
Required. A data frame containing at least circumplex scales.

scales
Required. The variable names or column numbers for the variables in .data that contain circumplex scales to be analyzed.

angles
Required. A numeric vector containing the angular displacement of each circumplex scale included in scales (in degrees).

measures
Optional. The variable names or column numbers for one or more variables in .data to be correlated with the circumplex scales and analyzed using correlation-based SSM analyses. To analyze the circumplex scales using mean-based analyses, simply omit this argument or set it to NULL (default = NULL).

grouping
Optional. The variable name or column number for the variable in .data that indicates the group membership of each observation. To analyze all observations in a single group, simply omit this argument or set it to NULL (default = NULL).

contrast
Optional. A string indicating what type of contrast to run. Current options are "none" for no contrast, "model" to find SSM parameters for the difference scores, or "test" to find the difference between the SSM parameters. Note that only two groups or measures can be contrasted at a time (default = "none").

boots
Optional. A single positive integer indicating how many bootstrap resamples to use when estimating the confidence intervals (default = 2000).

interval
Optional. A single positive number between 0 and 1 (exclusive) that indicates what confidence level to use when estimating the confidence intervals (default = 0.95).

listwise
Optional. A logical indicating whether missing values should be handled by listwise deletion (TRUE) or pairwise deletion (FALSE). Note that pairwise deletion may result in different missing data patterns in each bootstrap resample and is slower to compute (default = TRUE).

Value

A list containing the results and description of the analysis.

results

A tibble with the SSM parameter estimates

details

A list with the number of bootstrap resamples (boots), the confidence interval percentage level (interval), and the angular displacement of scales (angles)

call

A language object containing the function call that created this object

scores

A tibble containing the mean scale scores

type

A string indicating what type of SSM analysis was done

See also

Examples

# Load example data
data("jz2017")

# Single-group mean-based SSM
ssm_analyze(jz2017, scales = PA:NO, angles = octants())

#> Call:
#> ssm_analyze(.data = jz2017, scales = PA:NO, angles = octants())
#> 
#> Profile [All]:
#>                Estimate   Lower CI   Upper CI
#> Elevation         0.917      0.889      0.946
#> X-Value           0.351      0.326      0.379
#> Y-Value          -0.252     -0.280     -0.222
#> Amplitude         0.432      0.405      0.462
#> Displacement    324.292    321.063    327.900
#> Model Fit         0.878                      
#> 

# Single-group correlation-based SSM
ssm_analyze(jz2017, scales = PA:NO, angles = octants(),
  measures = c(NARPD, ASPD))

#> Call:
#> ssm_analyze(.data = jz2017, scales = PA:NO, angles = octants(), 
#>     measures = c(NARPD, ASPD))
#> 
#> Profile [NARPD]:
#>                Estimate   Lower CI   Upper CI
#> Elevation         0.202      0.171      0.235
#> X-Value          -0.062     -0.095     -0.028
#> Y-Value           0.179      0.144      0.213
#> Amplitude         0.189      0.153      0.225
#> Displacement    108.967     99.207    118.622
#> Model Fit         0.957                      
#> 
#> Profile [ASPD]:
#>                Estimate   Lower CI   Upper CI
#> Elevation         0.124      0.089      0.159
#> X-Value          -0.099     -0.132     -0.063
#> Y-Value           0.203      0.168      0.240
#> Amplitude         0.226      0.190      0.264
#> Displacement    115.927    107.038    124.248
#> Model Fit         0.964                      
#> 

  

# NOT RUN {
# Multiple-group mean-based SSM
ssm_analyze(jz2017, scales = PA:NO, angles = octants(), grouping = Gender)

# Multiple-group mean-based SSM with contrast
ssm_analyze(jz2017, scales = PA:NO, angles = octants(), grouping = Gender,
  contrast = "model")

# Single-group correlation-based SSM with contrast
ssm_analyze(jz2017, scales = PA:NO, angles = octants(),
  measures = c(NARPD, ASPD), contrast = "test")

# Multiple-group correlation-based SSM
ssm_analyze(jz2017, scales = PA:NO, angles = octants(), measures = NARPD,
  grouping = Gender)

# Multiple-group correlation-based SSM with contrast
ssm_analyze(jz2017, scales = PA:NO, angles = octants(), measures = NARPD,
  grouping = Gender, contrast = "test")
# }