Calculate Structural Summary Method parameters for a set of scores

Calculate SSM parameters (without confidence intervals) for a set of scores and generate a data frame with customizable labels for each parameter value. This function requires the input to be a numeric vector (or coercable to one) and returns only the parameters. See ssm_score() for a similar function that calculates SSM parameters for each row of a data frame.

Usage

ssm_parameters(
  scores,
  angles = octants(),
  prefix = "",
  suffix = "",
  e_label = "Elev",
  x_label = "Xval",
  y_label = "Yval",
  a_label = "Ampl",
  d_label = "Disp",
  f_label = "Fit"
)

Arguments

scores

Required. A numeric vector (or single row data frame) containing one score for each of a set of circumplex scales.

angles

Required. A numeric vector containing the angular displacement of each circumplex scale included in scores (in degrees).

prefix

Optional. A string to append to the beginning of all of the SSM parameters' variable names (default = "").

suffix

Optional. A string to append to the end of all of the SSM parameters' variable names (default = "").

e_label

Optional. A string representing the variable name of the SSM elevation parameter (default = "Elev").

x_label

Optional. A string representing the variable name of the SSM x-value parameter (default = "Xval").

y_label

Optional. A string representing the variable name of the SSM y-value parameter (default = "Yval").

a_label

Optional. A string representing the variable name of the SSM amplitude parameter (default = "Ampl").

d_label

Optional. A string representing the variable name of the SSM displacement parameter (default = "Disp").

f_label

Optional. A string representing the variable name of the SSM fit or R-squared value (default = "Fit").

Value

A data frame containing the SSM parameters calculated from scores.

See also

Examples

# Manually enter octant scores
scores <- c(0.55, 0.58, 0.62, 0.76, 1.21, 1.21, 1.48, 0.90)
ssm_parameters(scores)
#>      Elev      Xval       Yval      Ampl     Disp       Fit
#> 1 0.91375 0.3511181 -0.2516206 0.4319685 324.3736 0.8781155

# Customize several of the labels
ssm_parameters(scores, x_label = "LOV", y_label = "DOM")
#>      Elev       LOV        DOM      Ampl     Disp       Fit
#> 1 0.91375 0.3511181 -0.2516206 0.4319685 324.3736 0.8781155

# Add a prefix to all labels
ssm_parameters(scores, prefix = "IIP_")
#>   IIP_Elev  IIP_Xval   IIP_Yval  IIP_Ampl IIP_Disp   IIP_Fit
#> 1  0.91375 0.3511181 -0.2516206 0.4319685 324.3736 0.8781155